boot: split sealing tests and move some to fdestate's backend
